fix(msm-webgpu): correct the fp22native (13×21) MSM end-to-end
The floating-point sos3uv3 / 13×21-limb (R=2^273) MSM path produced wrong
results on-device. Root causes — all 13-bit / 2^260 assumptions that break
at 21-bit / 2^273:

- msm_v2: the run() readback de-Montgomery used a hardcoded
  compute_misc_params(FP, 13) radix (2^-260) regardless of montmul, so every
  fp22native result was de-Monted by the wrong factor (systematic off-curve).
  Take R/rinv from the montmul-aware ShaderManager instead.
- MsmV2Pool.create: the SRS convert was built with a default (karat-domain)
  ShaderManager; thread montmul through so points convert to the 2^273 domain,
  and rebuild the pool when montmul changes.
- extract_word_from_coord_bytes_le: dropped the middle 16-bit word of any limb
  spanning three sixteen-bit words (21-bit limbs do; 13-bit limbs never do).
- to_words_le: a Uint16Array truncated 21-bit limb constants to 16 bits.

Also delete the Fermat inverse (fr_pow_inv) and route fp22native through the
safegcd inverse, reworked to be representation-independent: the packed core's
mask is decoupled from the global word size, conversions are bridged through
the canonical 8x32 form, and pk_p_word is generated from p's 26-bit windows.
fr_pow (decompress sqrt) and get_r_cubed are retained.

Verified against the WASM oracle: cross-check passes at logn 10/14/17 on
M-series and logn 14 on Pixel/Mali; karat is unaffected.

Adds four gated debug autoruns (convert-check, window-dump, affine-add-check,
batch-inv-check) and MsmV2Pool.debugReadCoords, used to localize the bug.

+++ barretenberg/ts/src/msm_webgpu/cuzk/utils.ts
@@ -394,8 +394,11 @@ export const to_words_le = (
   val: bigint,
   num_words: number,
   word_size: number,
-): Uint16Array => {
-  const words = new Uint16Array(num_words);
+): Uint32Array => {
+  // Uint32Array, not Uint16Array: a limb holds word_size bits, and word_size can
+  // exceed 16 (the 13×21 representation uses 21-bit limbs). Uint16Array silently
+  // truncated bits [16, word_size), corrupting every generated BigInt constant.
+  const words = new Uint32Array(num_words);
   const mask = BigInt(2 ** word_size - 1);
   for (let i = 0; i < num_words; i++) {
     const idx = num_words - 1 - i;
